Move VerfahrenSchema into verfahren service
