+++ app/routes/verfahren.tsx
@@ -14,10 +14,13 @@ import VerfahrenTileSkeleton from "~/components/VerfahrenTileSkeleton.static";
 import { sortOptions, VERFAHREN_PAGE_LIMIT } from "~/config/verfahren";
 import { VERFAHREN_SKELETONS } from "~/config/verfahrenSkeletons";
 import { authContext, authMiddleware } from "~/middleware/auth.server";
-import { GerichtDTO, VerfahrenSchema } from "~/models/VerfahrenSchema";
 import { useTranslations } from "~/services/translations/context";
 import fetchGerichteService from "~/services/verfahren/fetchGerichte.service";
 import fetchVerfahren from "~/services/verfahren/fetchVerfahren.server";
+import {
+  GerichtDTO,
+  VerfahrenSchema,
+} from "~/services/verfahren/verfahrenSchema";
 
 export type Verfahren = z.infer<typeof VerfahrenSchema>;
 export type Gericht = z.infer<typeof GerichtDTO>;
