+++ app/services/verfahren/fetchGerichte.service.ts
@@ -1,8 +1,8 @@
 import z from "zod";
 import { serverConfig } from "~/config/config.server";
-import { CodeWertSchema } from "~/models/VerfahrenSchema";
 import { getBearerToken } from "~/services/auth/getBearerToken.server";
 import type { AuthenticationResponse } from "~/services/auth/oAuth.server";
+import { CodeWertSchema } from "./verfahrenSchema";
 
 const errorMessage =
   "Die Daten für das ausgewählte Gericht konnten nicht abgerufen werden.";
