+++ app/services/verfahren/fetchVerfahren.server.ts
@@ -1,10 +1,10 @@
 import z from "zod";
 import { serverConfig } from "~/config/config.server";
 import { sortOptions } from "~/config/verfahren";
-import { VerfahrenSchema } from "~/models/VerfahrenSchema";
 import { getBearerToken } from "~/services/auth/getBearerToken.server";
 import type { AuthenticationResponse } from "~/services/auth/oAuth.server";
 import { buildSearchParams } from "~/util/buildSearchParams";
+import { VerfahrenSchema } from "./verfahrenSchema";
 
 const fetchVerfahrenOptionsSchema = z.object({
   offset: z.number().int().nonnegative().optional(),
